requirements.txt
Flask
Flask-Login
Flask-WTF
SQLAlchemy
cryptography
Pillow
piexif
Werkzeug
flask-socketio 
